ref: merge backends x/n (#3477)
